+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- [#5] Refactor and split the dogu registry in a dogu version registry and a dogu spec repository.
 
 ## [v0.2.2] - 2024-08-02
 ### Added